fix(#1476): address review-first findings B1/B2/N1 on streaming do_get
Fix round on the streaming-do-get branch, addressing rust-reviewer
(APPROVE-with-Importants) + roborev (FAIL, 2 Medium) findings.

B1 (blocker): a panic in the blocking merge task dropped `tx`, the receiver
saw a normal channel close, and the RPC ended as a clean OK-EOF with a
silently TRUNCATED result. Fixed via `run_merge_catching_panics`
(std::panic::catch_unwind + AssertUnwindSafe) wrapping the merge body;
a caught panic forwards a new `ProducerError::Panicked { message }` into the
channel, mapped to `Status::internal` like any other producer error — so a
panic now surfaces as a mid-stream Status on every profile, not just under
panic=abort. New tests: panicking_merge_forwards_a_terminal_error_not_silent_close,
do_get_stream_surfaces_panic_as_internal_status_not_eof.

B2 (blocker): mid-stream errors no longer hit crate::obs::record_status_error
(pre-change do_get ran it for every error). Fixed by calling it in
MeteredDoGetStream's error arm before yielding Err. Verified via a
StreamProbe.errors_recorded counter (obs::record_error is a genuine no-op
without the `observability` feature, so a counter local to the crate is the
only portable way to prove the call site is reached) in the same panic test.

N1 (nit): reconciled the in-flight allowance into a single named
IN_FLIGHT_ALLOWANCE const (test-only) referenced by both the doc comment and
the test assertions, so they cannot drift.

File-size follow-up: the new tests pushed streaming.rs to 954 lines (over the
800 threshold it was previously under). Split its #[cfg(test)] mod tests out
into a sibling streaming_tests.rs via `#[path = "streaming_tests.rs"] mod
tests;` — streaming.rs is now 380 lines, streaming_tests.rs 575 (test-file
threshold 1500). producer.rs/service.rs remain over-threshold from before this
change (out of scope, #1116/#1135); lite re-run with
CQLITE_ALLOW_FILE_GROWTH=1 for those two only.

cargo test -p cqlite-flight: 147 passed. clippy -D warnings clean.
